refactor: convert StatusMessages and XBlockStatus components to typescript
This was requried to fix type errors raise in section and subsection
card components
