+++ src/course-unit/xblock-container-iframe/types.ts
@@ -1,16 +1,4 @@
-export interface GroupTypes {
-  id: number;
-  name: string;
-  selected: boolean;
-  deleted: boolean;
-}
-
-export interface UserPartitionTypes {
-  id: number;
-  name: string;
-  scheme: string;
-  groups: Array<GroupTypes>;
-}
+import { UserPartitionInfoTypes, UserPartitionTypes, XBlockPrereqs } from "../../data/types";
 
 export interface XBlockActionsTypes {
   canCopy: boolean;
@@ -50,27 +38,6 @@ export interface XBlockContainerIframeProps {
   handleConfigureSubmit: (XBlockId: string, ...args: any[]) => void;
 }
 
-export type UserPartitionInfoTypes = {
-  selectablePartitions: Array<{
-    groups: Array<{
-      deleted: boolean;
-      id: number;
-      name: string;
-      selected: boolean;
-    }>;
-    id: number;
-    name: string;
-    scheme: string;
-  }>;
-  selectedPartitionIndex: number;
-  selectedGroupsLabel: string;
-};
-
-export type PrereqTypes = {
-  blockDisplayName: string;
-  blockUsageKey: string;
-};
-
 export type AccessManagedXBlockDataTypes = {
   id: string;
   displayName?: string;
@@ -88,7 +55,7 @@ export type AccessManagedXBlockDataTypes = {
   userPartitionInfo?: UserPartitionInfoTypes;
   ancestorHasStaffLock?: boolean;
   isPrereq?: boolean;
-  prereqs?: PrereqTypes[];
+  prereqs?: XBlockPrereqs[];
   prereq?: number;
   prereqMinScore?: number;
   prereqMinCompletion?: number;
